+++ applications/effectMappingForm.js
@@ -1,5 +1,5 @@
 import { showArtSelect } from '../token-variants.mjs';
-import { SEARCH_TYPE, getFileName, isVideo, EXPRESSION_OPERATORS, keyPressed, FAUX_DOT } from '../scripts/utils.js';
+import { SEARCH_TYPE, getFileName, isVideo, keyPressed, FAUX_DOT } from '../scripts/utils.js';
 import TokenCustomConfig from './tokenCustomConfig.js';
 import { TVA_CONFIG, updateSettings } from '../scripts/settings.js';
 import EditJsonConfig from './configJsonEdit.js';
@@ -844,7 +844,7 @@ export default class EffectMappingForm extends FormApplication {
       m2.imgSrc = m1.imgSrc;
       m2.imgName = m1.imgName;
       m2.priority = m1.priority;
-      m2.effectName = m1.effectName.replaceAll('.', FAUX_DOT);
+      m2.effectName = m1.effectName.replaceAll('.', FAUX_DOT).replaceAll(String.fromCharCode(160), ' ');
       m2.overlay = m1.overlay;
       m2.alwaysOn = m1.alwaysOn;
       m2.disabled = m1.disabled;
@@ -857,15 +857,14 @@ export default class EffectMappingForm extends FormApplication {
 function highlightOperators(text) {
   text = text.replaceAll(FAUX_DOT, '.');
 
-  for (const op of EXPRESSION_OPERATORS) {
-    text = text.replaceAll(op, `<span>${op}</span>`);
-  }
-
   const re = new RegExp('([a-zA-Z\\.\\-\\|]+)([><=]+)(".*"|\\d+)(%{0,1})', `gi`);
   text = text.replace(re, function replace(match) {
     return '<span class="hp-expression">' + match + '</span>';
   });
 
+  for (const op of ['\\(', '\\)', '&&', '||', '\\!', '\\*', '\\{', '\\}']) {
+    text = text.replaceAll(op, `<span>${op}</span>`);
+  }
   return text;
 }